requirements.txt

Flask==3.0.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
Flask==2.3.3
Flask-SocketIO==5.3.6
pandas==2.1.1
numpy==1.25.2
scikit-learn==1.3.0
python-dotenv==1.0.0
APScheduler==3.10.4
schedule==1.2.0
plotly==5.17.0
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==20.6
smtplib-ssl==1.0.4
sqlite3
warnings
threading
json
datetime
os
time
